"clean"

+++ libasm.h
@@ -1,7 +1,7 @@
 #pragma once
 # include <stddef.h>
 
-extern int ft_strlen(char *str);
+extern size_t ft_strlen(char *str);
 extern int ft_strcmp(char *left, char *right);
 extern char *ft_strcpy(char *dst, char *src);
 extern size_t ft_write(int fd, const void *buf, size_t count);
